project/requirements.txt
# Streaming Voice Bot Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
# Note: audioop is built-in for Python 3.12 and earlier
# For Python 3.13+, install audioop-lts separately
